Move VaultManager to Service
